fix: utils namespace temporary read utils/index
I also removed docs for now. It was kind of an accident to remove it, but I plan on just grabbing it back from main and reintroducing it again. It kept having issues, but from what I've noticed, that the main package has the same issues when trying to run dev...